Fix Health Boost Health Duplication
This fixed the health duplication for the health boost potion effect.
Because the health boost effect adds to your max health it will get detected in the player.getMaxHealth() method.
When a player was to combat log with the effect it would duplicate their health every time because of this.